refactor: :truck: `회사소개` 하위 메뉴를 `회사정보`로 변경
페이지 내용은 수정하지 않음
