+++ frontend/src/App.tsx
@@ -5,7 +5,7 @@ import './App.css';
 import { HashRouter, Routes, Route } from "react-router-dom";
 import { Navigation } from './components'
 import { Main } from './pages'
-import { Vison, 회사소개 } from './pages/회사소개'
+import { Vison, 회사정보 } from './pages/회사소개'
 import { ESG경영 } from './pages/ESG'
 import { CareerInfo } from './pages/인재채용'
 import { 주신진엔지니어링, 서진티앤에이주 } from './pages/관계사현황'
@@ -18,8 +18,8 @@ function App() {
       <Routes>
         <Route path="/" element={<Main />} />
 
-        <Route path="/회사소개" element={<회사소개 />} />
-        <Route path="/회사소개/회사소개" element={<회사소개 />} />
+        <Route path="/회사소개" element={<회사정보 />} />
+        <Route path="/회사소개/회사정보" element={<회사정보 />} />
         <Route path="/회사소개/vison" element={<Vison />} />
 
         <Route path="/관계사현황" element={<주신진엔지니어링 />} />
