+++ frontend/src/pages/회사소개/index.tsx
@@ -1,4 +1,4 @@
 export { default as Vison } from "./Vison";
 export { default as 연혁 } from "./연혁";
 export { default as 조직도 } from "./조직도";
-export { default as 회사소개 } from "./회사소개";
+export { default as 회사정보 } from "./회사정보";
