+++ frontend/src/pages/회사소개/회사정보.css
